More optimization on the script
The remote dependencies must be a tar.gz so we can extract every deb
Check if the gh-pages remote branch exists before cloning it

+++ drone.io/build.sh
@@ -10,19 +10,29 @@ fi
 ## Include the repo configuration
 source packaging/drone.io/${REPO}.conf
 
-# Add our local repo
-mkdir repo
-cd repo
-wget ${REMOTE_DEPENDENCIES}
-sudo dpkg-scanpackages . /dev/null | gzip -c9 > Packages.gz
-sudo dpkg-scansources . | gzip -c9 > Sources.gz
-cd ..
-sudo apt-add-repository "deb file://${DRONE_BUILD_DIR}/repo /"
-sudo apt-get update
+## Add our local repo and our specific dependencies
+if [ ! -z ${REMOTE_DEPENDENCIES} ]; then
+	mkdir repo
+	cd repo
+	for i in ${REMOTE_DEPENDENCIES}; do
+		wget ${REMOTE_DEPENDENCIES} -O - | tar x
+	done
+	sudo dpkg-scanpackages . /dev/null | gzip -c9 > Packages.gz
+	sudo dpkg-scansources . | gzip -c9 > Sources.gz
+	cd ..
+	sudo apt-add-repository "deb file://${DRONE_BUILD_DIR}/repo /"
+	sudo apt-get update
+fi
 
 ## Install the dependencies
-sudo apt-get install devscripts cdbs check doxygen
-sudo apt-get install ${BUILD_DEPENDENCIES}
+sudo apt-get install devscripts cdbs check
+if [ ! -z ${BUILD_DEPENDENCIES} ]; then
+	sudo apt-get install ${BUILD_DEPENDENCIES}
+fi
+
+if [ -z ${NODOC} ]; then
+	sudo apt-get install doxygen
+fi
 
 ## Check that everything is fine
 NOCONFIGURE=1 ./autogen.sh
@@ -38,21 +48,25 @@ if [ -z ${NODEB} ]; then
 	if [ -e packaging/debian/${REPO}/debian ]; then
 		cp -r packaging/debian/${REPO}/debian .
 		debuild -i -us -uc -b
-		cp ../*.deb .
+		tar cf ${REPO}-debs.tar.gz ../*.deb
 	fi
 fi
 
 ## Push the doc to the gh-pages branch and docs directory
 ## Be sure to add the drone.io key to the deploy keys
 if [ -z ${NODOC} ]; then
-	git clone -b gh-pages git@github.com:turran/${REPO}.git gh-pages
-	make doc
-	rm -rf gh-pages/docs/*
-	cp -r doc/html/* gh-pages/docs
-	## Finally add the new files, remove the old ones, etc
-	cd gh-pages
-	git add -A
-	git commit -m "Updating the docs"
-	git push
-	cd ..
+	git branch -r | grep gh-pages
+	exist=$?
+	if [ ${exist} -eq 0 ]; then
+		git clone -b gh-pages git@github.com:turran/${REPO}.git gh-pages
+		make doc
+		rm -rf gh-pages/docs/*
+		cp -r doc/html/* gh-pages/docs
+		## Finally add the new files, remove the old ones, etc
+		cd gh-pages
+		git add -A
+		git commit -m "Updating the docs"
+		git push
+		cd ..
+	fi
 fi
